==> notebooks/04 - Split data into train and test .ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split

df = pd.read_csv("../data/df_clean.csv", low_memory=False, index_col=0)

df.sample(5)

test_rate = 0.1

df_train, df_test = train_test_split(df, test_size=test_rate, random_state=2907)

print("df_train shape", df_train.shape)
df_train = df_train.sort_index()
df_train.head(5)

print("df_test shape", df_test.shape)
df_test = df_test.sort_index()
df_test.head(5)

df_train.to_csv("../data/train.csv", index=False)
df_test.to_csv("../data/test.csv", index=False)
