==> notebooks/05 - Data preprocessing.ipynb <==
import pandas as pd
from sklearn.preprocessing import StandardScaler
import pickle

df = pd.read_csv("../data/train.csv", low_memory=False)
df.head(5)

X_scaler = StandardScaler()
y_scaler = StandardScaler()

X = df.drop(["smiles", "homolumogap"], axis=1).to_numpy()
y = df[["homolumogap"]].to_numpy()

X_scaled = X_scaler.fit_transform(X)
y_scaled = y_scaler.fit_transform(y)

with open("../scalers/X_scaler.pkl", "wb") as f:
    pickle.dump(X_scaler, f)

with open("../scalers/y_scaler.pkl", "wb") as f:
    pickle.dump(y_scaler, f)

X_scaled.dump("../data/X_train_scaled.npy")
y_scaled.dump("../data/y_train_scaled.npy")
