==> notebooks/07.5 - Non Linear Regression with Neural Networks.ipynb <==
import numpy as np

from tensorflow.keras.layers import (
    Dense,
    Dropout,
)
from tensorflow.keras.models import Sequential
from scikeras.wrappers import KerasRegressor
from sklearn.model_selection import GridSearchCV
from tensorflow.keras.callbacks import ReduceLROnPlateau
from sklearn.metrics import r2_score, mean_squared_error
import pickle

X = np.load("../data/X_train_scaled.npy", allow_pickle=True)
y = np.load("../data/y_train_scaled.npy", allow_pickle=True)

print("X shape:", X.shape)
print("y shape:", y.shape)

lr_reduce = ReduceLROnPlateau(
    min_delta=1e-5,
    patience=5,
    verbose=1
)

callbacks = [
    lr_reduce
]

def create_model(n_neurons, input_shape=(X.shape[1], )):
    model = Sequential()
    model.add(Dense(n_neurons, activation="relu", input_shape=input_shape))
    model.add(Dropout(0.3))
    model.add(Dense(n_neurons, activation="relu"))
    model.add(Dropout(0.3))
    model.add(Dense(1, activation="linear"))
    model.compile(optimizer="adam", loss="mse")
    return model

regressor = KerasRegressor(
    model=create_model,
    n_neurons=32,
    batch_size=128,
    epochs=100,
)

param_grid = {
    "n_neurons": [32, 64, 128],
}

grid_search = GridSearchCV(
    regressor,
    param_grid,
    scoring="neg_mean_squared_error",
    cv=10,
    verbose=10
)

np.random.seed(2907)
grid_search.fit(X, y)

print("GridSearch results:")
grid_search.cv_results_

with open("../grids/grid_search_neural_net_01.pkl", mode="wb") as f:
    pickle.dump(grid_search, f)

best_regressor = grid_search.best_estimator_

print("Best regressor", best_regressor)

with open("../models/model_gradient_neural_net_01.pkl", mode="wb") as f:
    pickle.dump(best_regressor, f)

y_hat = best_regressor.predict(X)

r2 = r2_score(y, y_hat)
mse = mean_squared_error(y, y_hat)

print(f"R2 Score: {r2:.3f}")
print(f"MSE: {mse:.3f}")
