==> notebooks/08 - Assessing model performances.ipynb <==
## Imports

import os
import pickle

from tqdm.notebook import tqdm
import numpy as np
import pandas as pd
pd.options.display.width = 0

from tensorflow.keras.layers import (
    Dense,
    Dropout,
)
from tensorflow.keras.models import Sequential

from sklearn.metrics import mean_squared_error, mean_absolute_error

## Grid search models comparison

### Import dumped grids

grids = [grid for grid in os.listdir('../grids') if 'grid_search' in grid]
grids

### Load other necessary files

X = np.load("../data/X_train_scaled.npy", allow_pickle=True)
y = np.load("../data/y_train_scaled.npy", allow_pickle=True)

def create_model(n_neurons, dropout_rate, input_shape=(X.shape[1], )):
    model = Sequential()
    model.add(Dense(n_neurons, activation="relu", input_shape=input_shape))
    if dropout_rate > 0:
        model.add(Dropout(dropout_rate))
    model.add(Dense(n_neurons, activation="relu"))
    if dropout_rate > 0:
        model.add(Dropout(dropout_rate))
    model.add(Dense(1, activation="linear"))
    model.compile(optimizer="adam", loss="mse")
    return model

### Create comparison DataFrame

df_grid = pd.DataFrame()

for grid in grids:
    model_type = grid[grid.index("_", 8)+1:grid.rindex("_")]
    with open(f"../grids/{grid}", mode="rb") as f:
        pkl = pickle.load(f)
        gdf = pd.DataFrame.from_dict(pkl.cv_results_)[
            ["params", "mean_test_score", "std_test_score"]
        ]
        gdf["model_type"] = model_type
        df_grid = df_grid.append(
            gdf,
            ignore_index=True
        )

df_grid = df_grid.sort_values(
    "mean_test_score",
    ascending=False
).reset_index(drop=True)

print("All models")
df_grid

print("Best models")
df_grid.drop_duplicates('model_type')

with open("../scalers/X_scaler.pkl", "rb") as f:
    X_scaler = pickle.load(f)

with open("../scalers/y_scaler.pkl", "rb") as f:
    y_scaler = pickle.load(f)

df_test = pd.read_csv("../data/test.csv")

X_test = X_scaler.transform(
    df_test.drop(["smiles", "homolumogap"], axis=1).to_numpy()
)
y_test = y_scaler.transform(
    df_test[["homolumogap"]].to_numpy()
)

df_test_result = pd.DataFrame(columns=["model_type", "mse", "mae"])

pbar = tqdm(grids)

for grid in pbar:
    model_type = grid[grid.index("_", 8)+1:grid.rindex("_")]
    pbar.set_description(f"Predicting with {model_type}...")
    
    with open(f"../grids/{grid}", mode="rb") as f:
        pkl = pickle.load(f)
    model = pkl.best_estimator_
    y_hat = model.predict(X_test)
    mse = mean_squared_error(y_test, y_hat)
    mae = mean_absolute_error(y_test, y_hat)
    df_test_result = df_test_result.append(
        {
            "model_type": model_type,
            "mse": mse,
            "mae": mae,
        },
        ignore_index=True
    )

df_test_result.sort_values(["mse", "mae"]).reset_index(drop=True)
